requirements.txt
fastapi[standard]
uvicorn[standard]
transformers
torch
pymupdf
pdfminer.six
python-multipart
huggingface-hub
pydantic
pytesseract
pillow
